backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.23,<1.5.0
pydantic>=1.8.0,<2.0.0
python-dotenv>=0.19.0,<0.20.0
python-telegram-bot>=20.0.0
